testharness: build out the Context concept
Contexts now also have names, and an interface called contextSetter can
be used to derive new contexts from parent contexts. contextSetter will
be the type of the first argument to the 'describe' function in a test
spec script, which can either be a simple string (turned into a
simpleContextSetter) or a userdata returned from a function that wraps
one of the other contextSetter implementations.